split repl_commands.rs to reduce file size
